refactor(rich-text-editor): rename .scss files to .module.scss
