+++ packages/rich-text-editor/src/EditableRichTextContent/EditableRichTextContent.module.scss
