+++ packages/rich-text-editor/src/RichTextContent/RichTextContent.module.scss
