+++ packages/rich-text-editor/src/RichTextEditor/RichTextEditor.module.scss
