+++ packages/rich-text-editor/src/RichTextEditor/components/Toolbar/Toolbar.module.scss
