+++ packages/rich-text-editor/src/RichTextEditor/components/ToolbarSection/ToolbarSection.module.scss
